Version update to v0.9.6
Added classes such as MessageDialog, Sliders, and threads to help build
loading screens

+++ sky/engine/physics/bounding/BoundingAABB.java
@@ -155,7 +155,7 @@ public class BoundingAABB extends Bounding
 	 * Rotate is unsupported
 	 */
 	@Override
-	public void rotate(int degree)
+	public void rotate(float degree)
 	{
 		return;
 	}
